# config/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "identity-verificate-service",
        "version": "1.0",
        "description": "本人性確認サービス"
    },
    "servers": [
        {
            "url": "/identity-verificate"
        }
    ],
    "paths": {
...
        "/url": {
            "post": {
                "tags": [
                    "URL発行"
                ],
                "description": "本人性確認を閲覧する為のURLを発行します。",
                "requestBody": {
                    "description": "発行するURL情報",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "発行するURLに関するリクエスト",
                                "properties": {
                                    "identifyCode": {
                                        "type": "string",
                                        "description": "本人性確認コード"
                                    },
                                    "urlType": {
                                        "type": "number",
                                        "required": [
                                            "0",
                                            "1"
                                        ],
                                        "description": "発行URL種別(0: 本人用, 1: 第三者用)"
                                    },
                                    "userId": {
                                        "type": "string",
                                        "description": "ユーザーID"
                                    }
                                },
                                "example": {
                                    "identifyCode": "ukO8z+Xf8vv7yxXQj2Hpo",
                                    "urlType": 1,
                                    "userId": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "URLの発行に成功した際のレスポンス",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "発行URL",
                                    "properties": {
                                        "url": {
                                            "type": "string",
                                            "description": "発行URL"
                                        }
                                    },
                                    "example": {
                                        "url": "https://xxxx.xxx/confirm/third-party/{本人性確認コード}/"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/url/issue": {
            "post": {
                "tags": [
                    "URL発行"
                ],
                "description": "本人性確認を閲覧する為のURLを発行します。",
                "requestBody": {
                    "description": "発行するURL情報",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "発行するURLに関するリクエスト",
                                "properties": {
                                    "identifyCode": {
                                        "type": "string",
                                        "description": "本人性確認コード"
                                    },
                                    "urlType": {
                                        "type": "number",
                                        "required": [
                                            "0",
                                            "1"
                                        ],
                                        "description": "発行URL種別(0: 本人用, 1: 第三者用)"
                                    },
                                    "userId": {
                                        "type": "string",
                                        "description": "ユーザーID"
                                    }
                                },
                                "example": {
                                    "urlType": 1,
                                    "userId": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "URLの発行に成功した際のレスポンス",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "発行URL",
                                    "properties": {
                                        "url": {
                                            "type": "string",
                                            "description": "発行URL"
                                        }
                                    },
                                    "example": {
                                        "url": "https://xxxx.xxx/confirm/third-party/{本人性確認コード}/"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/url/confirm": {
            "post": {
                "tags": [
                    "URL確認"
                ],
                "description": "本人性確認を閲覧する為のURLを発行します。",
                "requestBody": {
                    "description": "発行するURL情報",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "発行するURLに関するリクエスト",
                                "properties": {
                                    "path": {
                                        "type": "string",
                                        "description": "パス"
                                    }
                                },
                                "example": {
                                    "path": "OTY1MDZlYjktYmY5NC00NmQ1LWI1NzQtYTFiZDkxNzQzNzJj"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "URLの発行に成功した際のレスポンス",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "発行URL",
                                    "properties": {
                                        "url": {
                                            "type": "string",
                                            "description": "発行URL"
                                        }
                                    },
                                    "example": {
                                        "url": "https://xxxx.xxx/confirm/third-party/{本人性確認コード}/"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}